// apps/http/src/infrastructure/email/dtos/send-email.dto.ts
import { EmailPayload, EmailTemplateType } from "../entities/email.entity";

export interface EnqueueEmailInputDTO {
    templateType: EmailTemplateType;
    payload: EmailPayload;
    metadata?: Record<string, any>;
    priority?: number;
    delay?: number;
}

export interface EnqueueEmailOutputDTO {
    jobId: string;
    enqueued: boolean;
    timestamp: string;
}